refactor(front): move tables to views subdirectory
