+++ front/src/features/tables/views/EmptyLogs.tsx
